requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psutil==5.9.2
py-cpuinfo==8.0.0
pymongo==4.2.0
python-dotenv==0.21.0
-e git+https://github.com/jpseara/python3-nmap.git@76b76a5c988e5ad925f2cd6603af4ddf6ff09deb#egg=python3_nmap
scapy==2.4.5
simplejson==3.17.6
waitress==2.1.2
Werkzeug==2.2.2
zipp==3.9.0
